Fix up zoom action
Currently, zoom action doesn't work too well when you zoom in and out
of the sequence because of a percentage zoom factor. Switching to a set
value for zoom increment fixes this. Also, we set the zoom to a center
element now so that it doesn't scroll off content while zooming.